+++ axum-extra/src/response/erased_json.rs
@@ -1,5 +1,4 @@
 use axum::{
-    body::{self, Full},
     http::{header, HeaderValue, StatusCode},
     response::{IntoResponse, Response},
 };
@@ -49,22 +48,16 @@ impl ErasedJson {
 
 impl IntoResponse for ErasedJson {
     fn into_response(self) -> Response {
-        let bytes = match self.0 {
-            Ok(res) => res,
-            Err(err) => {
-                return Response::builder()
-                    .status(StatusCode::INTERNAL_SERVER_ERROR)
-                    .header(header::CONTENT_TYPE, mime::TEXT_PLAIN_UTF_8.as_ref())
-                    .body(body::boxed(Full::from(err.to_string())))
-                    .unwrap();
-            }
-        };
-
-        let mut res = Response::new(body::boxed(Full::new(bytes)));
-        res.headers_mut().insert(
-            header::CONTENT_TYPE,
-            HeaderValue::from_static(mime::APPLICATION_JSON.as_ref()),
-        );
-        res
+        match self.0 {
+            Ok(bytes) => (
+                [(
+                    header::CONTENT_TYPE,
+                    HeaderValue::from_static(mime::APPLICATION_JSON.as_ref()),
+                )],
+                bytes,
+            )
+                .into_response(),
+            Err(err) => (StatusCode::INTERNAL_SERVER_ERROR, err.to_string()).into_response(),
+        }
     }
 }