+++ axum/src/extract/extension.rs
@@ -1,5 +1,7 @@
 use super::{rejection::*, FromRequest, RequestParts};
+use crate::response::IntoResponseParts;
 use async_trait::async_trait;
+use axum_core::response::{IntoResponse, Response, ResponseParts};
 use std::ops::Deref;
 
 /// Extractor that gets a value from request extensions.
@@ -71,5 +73,25 @@ impl<T> Deref for Extension<T> {
 
     fn deref(&self) -> &Self::Target {
         &self.0
+    }
+}
+
+impl<T> IntoResponseParts for Extension<T>
+where
+    T: Send + Sync + 'static,
+{
+    fn into_response_parts(self, res: &mut ResponseParts) {
+        res.insert_extension(self.0);
+    }
+}
+
+impl<T> IntoResponse for Extension<T>
+where
+    T: Send + Sync + 'static,
+{
+    fn into_response(self) -> Response {
+        let mut res = ().into_response();
+        res.extensions_mut().insert(self.0);
+        res
     }
 }