+++ axum/src/json.rs
@@ -1,10 +1,10 @@
 use crate::{
-    body::{self, Bytes, Full, HttpBody},
+    body::{Bytes, HttpBody},
     extract::{rejection::*, FromRequest, RequestParts},
-    response::{IntoResponse, Response},
     BoxError,
 };
 use async_trait::async_trait;
+use axum_core::response::{IntoResponse, Response};
 use http::{
     header::{self, HeaderValue},
     StatusCode,
@@ -158,26 +158,25 @@ where
     T: Serialize,
 {
     fn into_response(self) -> Response {
-        let bytes = match serde_json::to_vec(&self.0) {
-            Ok(res) => res,
-            Err(err) => {
-                return Response::builder()
-                    .status(StatusCode::INTERNAL_SERVER_ERROR)
-                    .header(
-                        header::CONTENT_TYPE,
-                        HeaderValue::from_static(mime::TEXT_PLAIN_UTF_8.as_ref()),
-                    )
-                    .body(body::boxed(Full::from(err.to_string())))
-                    .unwrap();
-            }
-        };
-
-        let mut res = Response::new(body::boxed(Full::from(bytes)));
-        res.headers_mut().insert(
-            header::CONTENT_TYPE,
-            HeaderValue::from_static(mime::APPLICATION_JSON.as_ref()),
-        );
-        res
+        match serde_json::to_vec(&self.0) {
+            Ok(bytes) => (
+                [(
+                    header::CONTENT_TYPE,
+                    HeaderValue::from_static(mime::APPLICATION_JSON.as_ref()),
+                )],
+                bytes,
+            )
+                .into_response(),
+            Err(err) => (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                [(
+                    header::CONTENT_TYPE,
+                    HeaderValue::from_static(mime::TEXT_PLAIN_UTF_8.as_ref()),
+                )],
+                err.to_string(),
+            )
+                .into_response(),
+        }
     }
 }