+++ axum/src/response/mod.rs
@@ -1,10 +1,8 @@
 #![doc = include_str!("../docs/response.md")]
 
 use crate::body::{Bytes, Full};
-use axum_core::body::boxed;
 use http::{header, HeaderValue};
 
-mod headers;
 mod redirect;
 
 pub mod sse;
@@ -14,10 +12,10 @@ pub mod sse;
 pub use crate::Json;
 
 #[doc(inline)]
-pub use axum_core::response::{IntoResponse, IntoResponseHeaders, Response};
+pub use axum_core::response::{IntoResponse, IntoResponseParts, Response, ResponseParts};
 
 #[doc(inline)]
-pub use self::{headers::Headers, redirect::Redirect, sse::Sse};
+pub use self::{redirect::Redirect, sse::Sse};
 
 /// An HTML response.
 ///
@@ -30,12 +28,14 @@ where
     T: Into<Full<Bytes>>,
 {
     fn into_response(self) -> Response {
-        let mut res = Response::new(boxed(self.0.into()));
-        res.headers_mut().insert(
-            header::CONTENT_TYPE,
-            HeaderValue::from_static(mime::TEXT_HTML_UTF_8.as_ref()),
-        );
-        res
+        (
+            [(
+                header::CONTENT_TYPE,
+                HeaderValue::from_static(mime::TEXT_HTML_UTF_8.as_ref()),
+            )],
+            self.0.into(),
+        )
+            .into_response()
     }
 }
 
@@ -47,43 +47,166 @@ impl<T> From<T> for Html<T> {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::body::Empty;
-    use http::{
-        header::{HeaderMap, HeaderName},
-        StatusCode,
-    };
+    use crate::extract::Extension;
+    use crate::{body::Body, routing::get, Router};
+    use axum_core::response::IntoResponse;
+    use http::HeaderMap;
+    use http::{StatusCode, Uri};
 
-    #[test]
-    fn test_merge_headers() {
-        struct MyResponse;
+    // just needs to compile
+    #[allow(dead_code)]
+    fn impl_trait_result_works() {
+        async fn impl_trait_ok() -> Result<impl IntoResponse, ()> {
+            Ok(())
+        }
 
-        impl IntoResponse for MyResponse {
-            fn into_response(self) -> Response {
-                let mut resp = Response::new(boxed(Empty::new()));
-                resp.headers_mut()
-                    .insert(HeaderName::from_static("a"), HeaderValue::from_static("1"));
-                resp
+        async fn impl_trait_err() -> Result<(), impl IntoResponse> {
+            Err(())
+        }
+
+        async fn impl_trait_both(uri: Uri) -> Result<impl IntoResponse, impl IntoResponse> {
+            if uri.path() == "/" {
+                Ok(())
+            } else {
+                Err(())
             }
         }
 
-        fn check(resp: impl IntoResponse) {
-            let resp = resp.into_response();
-            assert_eq!(
-                resp.headers().get(HeaderName::from_static("a")).unwrap(),
-                &HeaderValue::from_static("1")
-            );
-            assert_eq!(
-                resp.headers().get(HeaderName::from_static("b")).unwrap(),
-                &HeaderValue::from_static("2")
-            );
+        async fn impl_trait(uri: Uri) -> impl IntoResponse {
+            if uri.path() == "/" {
+                Ok(())
+            } else {
+                Err(())
+            }
         }
 
-        let headers: HeaderMap =
-            std::iter::once((HeaderName::from_static("b"), HeaderValue::from_static("2")))
-                .collect();
+        Router::<Body>::new()
+            .route("/", get(impl_trait_ok))
+            .route("/", get(impl_trait_err))
+            .route("/", get(impl_trait_both))
+            .route("/", get(impl_trait));
+    }
 
-        check((headers.clone(), MyResponse));
-        check((StatusCode::OK, headers, MyResponse));
+    // just needs to compile
+    #[allow(dead_code)]
+    fn tuple_responses() {
+        async fn status() -> impl IntoResponse {
+            StatusCode::OK
+        }
+
+        async fn status_headermap() -> impl IntoResponse {
+            (StatusCode::OK, HeaderMap::new())
+        }
+
+        async fn status_header_array() -> impl IntoResponse {
+            (StatusCode::OK, [("content-type", "text/plain")])
+        }
+
+        async fn status_headermap_body() -> impl IntoResponse {
+            (StatusCode::OK, HeaderMap::new(), String::new())
+        }
+
+        async fn status_header_array_body() -> impl IntoResponse {
+            (
+                StatusCode::OK,
+                [("content-type", "text/plain")],
+                String::new(),
+            )
+        }
+
+        async fn status_headermap_impl_into_response() -> impl IntoResponse {
+            (StatusCode::OK, HeaderMap::new(), impl_into_response())
+        }
+
+        async fn status_header_array_impl_into_response() -> impl IntoResponse {
+            (
+                StatusCode::OK,
+                [("content-type", "text/plain")],
+                impl_into_response(),
+            )
+        }
+
+        fn impl_into_response() -> impl IntoResponse {}
+
+        async fn status_header_array_extension_body() -> impl IntoResponse {
+            (
+                StatusCode::OK,
+                [("content-type", "text/plain")],
+                Extension(1),
+                String::new(),
+            )
+        }
+
+        async fn status_header_array_extension_mixed_body() -> impl IntoResponse {
+            (
+                StatusCode::OK,
+                [("content-type", "text/plain")],
+                Extension(1),
+                HeaderMap::new(),
+                String::new(),
+            )
+        }
+
+        //
+
+        async fn headermap() -> impl IntoResponse {
+            HeaderMap::new()
+        }
+
+        async fn header_array() -> impl IntoResponse {
+            [("content-type", "text/plain")]
+        }
+
+        async fn headermap_body() -> impl IntoResponse {
+            (HeaderMap::new(), String::new())
+        }
+
+        async fn header_array_body() -> impl IntoResponse {
+            ([("content-type", "text/plain")], String::new())
+        }
+
+        async fn headermap_impl_into_response() -> impl IntoResponse {
+            (HeaderMap::new(), impl_into_response())
+        }
+
+        async fn header_array_impl_into_response() -> impl IntoResponse {
+            ([("content-type", "text/plain")], impl_into_response())
+        }
+
+        async fn header_array_extension_body() -> impl IntoResponse {
+            (
+                [("content-type", "text/plain")],
+                Extension(1),
+                String::new(),
+            )
+        }
+
+        async fn header_array_extension_mixed_body() -> impl IntoResponse {
+            (
+                [("content-type", "text/plain")],
+                Extension(1),
+                HeaderMap::new(),
+                String::new(),
+            )
+        }
+
+        Router::<Body>::new()
+            .route("/", get(status))
+            .route("/", get(status_headermap))
+            .route("/", get(status_header_array))
+            .route("/", get(status_headermap_body))
+            .route("/", get(status_header_array_body))
+            .route("/", get(status_headermap_impl_into_response))
+            .route("/", get(status_header_array_impl_into_response))
+            .route("/", get(status_header_array_extension_body))
+            .route("/", get(status_header_array_extension_mixed_body))
+            .route("/", get(headermap))
+            .route("/", get(header_array))
+            .route("/", get(headermap_body))
+            .route("/", get(header_array_body))
+            .route("/", get(headermap_impl_into_response))
+            .route("/", get(header_array_impl_into_response))
+            .route("/", get(header_array_extension_body))
+            .route("/", get(header_array_extension_mixed_body));
     }
 }