+++ axum/src/response/sse.rs
@@ -28,9 +28,12 @@
 //! ```
 
 use crate::{
-    body::{self, Bytes, HttpBody},
-    response::{IntoResponse, Response},
+    body::{Bytes, HttpBody},
     BoxError,
+};
+use axum_core::{
+    body,
+    response::{IntoResponse, Response},
 };
 use bytes::{BufMut, BytesMut};
 use futures_util::{
@@ -95,16 +98,17 @@ where
     E: Into<BoxError>,
 {
     fn into_response(self) -> Response {
-        let body = body::boxed(Body {
-            event_stream: SyncWrapper::new(self.stream),
-            keep_alive: self.keep_alive.map(KeepAliveStream::new),
-        });
-
-        Response::builder()
-            .header(http::header::CONTENT_TYPE, mime::TEXT_EVENT_STREAM.as_ref())
-            .header(http::header::CACHE_CONTROL, "no-cache")
-            .body(body)
-            .unwrap()
+        (
+            [
+                (http::header::CONTENT_TYPE, mime::TEXT_EVENT_STREAM.as_ref()),
+                (http::header::CACHE_CONTROL, "no-cache"),
+            ],
+            body::boxed(Body {
+                event_stream: SyncWrapper::new(self.stream),
+                keep_alive: self.keep_alive.map(KeepAliveStream::new),
+            }),
+        )
+            .into_response()
     }
 }